【修改】 统计加上上方的行情综合统计

+++ src/main/java/com/example/springboot/index/vo/vo/StatisticCommon.java
@@ -13,5 +13,5 @@ public class StatisticCommon {
     private String indexCodeL1;
     private String indexCodeL2;
     private String indexCodeL3;
-    private Integer totalMv;
+    private Double totalMv;
 }
